requirements.txt
torch==2.2.0
torchvision==0.17.0
opencv-python==4.10.0.84
pydicom==2.4.4
matplotlib==3.8.2
numpy==1.26.3
flask==3.0.0
flask-cors==4.0.0
pillow==10.2.0
gunicorn==21.2.0
tensorboard==2.15.1